yunseokgithub(0217)

+++ src/main/java/com/wangin/admin/dto/GolfDto.java
@@ -3,6 +3,8 @@ package com.wangin.admin.dto;
 import com.wangin.admin.entity.GolfEntity;
 import lombok.*;
 
+import java.util.Date;
+
 @Getter
 @Setter
 @ToString
@@ -11,11 +13,11 @@ public class GolfDto {
 
     private Long RI_NO;
 
-    private int RI_CA_NO;
+    private Long RI_CA_NO;
 
-    private int RI_UI_NO;
+    private Long RI_UI_NO;
 
-    private int RI_CC_NO;
+    private Long RI_CC_NO;
 
     private String RI_CA_ID;
 
@@ -51,7 +53,7 @@ public class GolfDto {
 
     @Builder
 
-    public GolfDto(Long RI_NO, int RI_CA_NO, int RI_UI_NO, int RI_CC_NO, String RI_CA_ID, String RI_CA_PASSWORD, int RI_YEAR, int RI_MONTH, int RI_DAY, int RI_START_TIME, int RI_END_TIME, int RI_HALL, int RI_COURSE, int RI_PERSON, int RI_SEX,
+    public GolfDto(Long RI_NO, Long RI_CA_NO, Long RI_UI_NO, Long RI_CC_NO, String RI_CA_ID, String RI_CA_PASSWORD, int RI_YEAR, int RI_MONTH, int RI_DAY, int RI_START_TIME, int RI_END_TIME, int RI_HALL, int RI_COURSE, int RI_PERSON, int RI_SEX,
                    int RI_STATE, int RI_CHOICE, String RI_CANCLE_DAY, String RI_IDATETIME, String RI_UDATETIME) {
         this.RI_NO = RI_NO;
         this.RI_CA_NO = RI_CA_NO;
